new: avoid reordering mapping keys in the output.
Several commands from ``shyaml`` will output YAML themselves. But in the
process, the whole content is parsed and re-dumped. This can lead to
differences with the original inputs. The ordering of keys can be
annoying in some use cases, even if, well mappings should not be
ordered, these are ordered by the sequential nature of a file, and using
'omap' is often tedious. ``shyaml`` now keep the original order the keys
were met.