new: introduce ``-y`` option to force full yaml output.
By default, output of literal string where directly printed unquoted and
useable. If this is very useful, there are many cases where this is not
consistent and is ambiguous, especially when this string is itself a
YAML representation.

This could actually lead to subtle bugs occuring when re-using
``shyaml`` output for further parsing.

So if you want unquoted literal types to further process these elements,
use ``-y``. If you expect a string and wan't it plain raw, avoid ``-y``.